refactor: SectionEdge는 값 객체가 아니라고 생각되어 수정하였습니다.
