+++ src/main/java/wooteco/subway/domain/SectionEdge.java
@@ -1,7 +1,6 @@
-package wooteco.subway.domain.vo;
+package wooteco.subway.domain;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
-import wooteco.subway.domain.Section;
 
 public class SectionEdge extends DefaultWeightedEdge {
